+++ changelog.md
@@ -7,6 +7,7 @@ Version 19.03
 + String `__version__` provided.
 + Triangulation data type changed for performance improvement.
 + Change "print" function as "logger.debug".
++ Embed the "Coordinate" object in return value to improve performance.
 
 Version 19.02
 ===